config/env.py
from dotenv import load_dotenv
import os

load_dotenv()  # loads the .env file

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

if not OPENAI_API_KEY:
    raise ValueError("OPENAI_API_KEY not set in environment variables")
